# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book organizer API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/user/": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "Get one user",
                "operationId": "0b0a3b2957d6f543e8cd6ec456a8f05b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Register a new user",
                "operationId": "645c07ec1432f0434d9e8f7af91626ec",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Created"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "7|aI87lhNyFBrWX73elXpYpDbzQ38YciDmN..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict"
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Log in ",
                "operationId": "fa14a7fd1f8095d2598c5f6a530327b6",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "7|aI87lhNyFBrWX73elXpYpDbzQ38YciDmN..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad credentials"
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Log out ",
                "operationId": "0eeae717bf70f0e5d23609dd16c495f4",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/user/edit": {
            "put": {
                "tags": [
                    "Users"
                ],
                "description": "Edit user data ",
                "operationId": "6c0b445d83cbd47c32a27b44d7ca77fb",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}